agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn",
    "pydantic-ai-slim[ag-ui,openai]",
    "python-dotenv",
    "logfire>=4.10.0",
    "fastapi>=0.128.2",
    "sqlmodel",
    "psycopg[binary]",
]
